+++ server/src/main/java/ru/practicum/shareit/request/repository/ItemRequestRepository.java
@@ -6,10 +6,10 @@ import ru.practicum.shareit.request.entity.ItemRequest;
 
 import java.util.List;
 
-public interface ItemRequestRepository extends JpaRepository<ItemRequest, Integer> {
+public interface ItemRequestRepository extends JpaRepository<ItemRequest, Long> {
     @Query("select i from ItemRequest i where i.requester.id = ?1")
-    List<ItemRequest> findByRequesterId(Integer requesterId);
+    List<ItemRequest> findByRequesterId(Long requesterId);
 
     @Query("select i from ItemRequest i where i.requester.id <> ?1")
-    List<ItemRequest> findByRequesterIdNot(Integer requesterId);//, Pageable pageable);
+    List<ItemRequest> findByRequesterIdNot(Long requesterId);//, Pageable pageable);
 }
